+++ routes/web.php
@@ -3,7 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use LaraZeus\Wind\Http\Livewire\Contacts;
 
-Route::middleware(config('zeus.middleware'))
-    ->prefix(config('zeus.prefix'))
-    ->get('contact-us/{department?}', Contacts::class)
+Route::middleware(config('zeus-wind.middleware'))
+    ->prefix(config('zeus-wind.prefix'))
+    ->get('contact-us/{departmentSlug?}', Contacts::class)
     ->name('contact');