+++ src/Filament/Resources/LetterResource/Pages/EditLetter.php
@@ -12,7 +12,7 @@ class EditLetter extends EditRecord
     public function mount($record): void
     {
         parent::mount($record);
-        if ($this->record->status === 'NEW') {
+        if ($this->record->status === config('zeus-wind.default_status')) {
             $this->record->update(['status' => 'READ']);
         }
     }